+++ src/main/java/commoble/hyperbox/client/BlockPreviewRenderer.java
@@ -15,12 +15,15 @@ import net.minecraft.client.renderer.block.ModelBlockRenderer;
 import net.minecraft.client.renderer.block.BlockRenderDispatcher;
 import net.minecraft.client.renderer.MultiBufferSource;
 import net.minecraft.client.renderer.RenderType;
+import net.minecraft.client.renderer.Sheets;
 import net.minecraft.client.renderer.block.model.BakedQuad;
 import net.minecraft.client.renderer.texture.OverlayTexture;
 import com.mojang.blaze3d.vertex.DefaultVertexFormat;
 import net.minecraft.core.BlockPos;
 import com.mojang.math.Matrix4f;
 import net.minecraft.world.phys.Vec3;
+import net.minecraftforge.client.model.data.ModelData;
+
 import com.mojang.math.Vector3f;
 import net.minecraft.core.Vec3i;
 import com.mojang.math.Vector4f;
@@ -64,100 +67,45 @@ public class BlockPreviewRenderer extends ModelBlockRenderer
 	
 		BlockRenderDispatcher blockDispatcher = Minecraft.getInstance().getBlockRenderer();
 		ModelBlockRenderer renderer = getInstance(blockDispatcher.getModelRenderer());
-		renderer.tesselateBlock(
+		// this render type is which buffer to render to
+		RenderType bufferType = Sheets.translucentCullBlockSheet();
+		// this render type is which render type to get quads for from the model
+		// null => all quads
+		RenderType renderType = null;
+		renderer.tesselateWithoutAO(
 			level,
 			blockDispatcher.getBlockModel(state),
 			state,
 			pos,
 			matrix,
-			renderTypeBuffer.getBuffer(RenderType.translucent()),
+			renderTypeBuffer.getBuffer(bufferType),
 			false,
 			level.random,
 			state.getSeed(pos),
 			OverlayTexture.NO_OVERLAY,
-			net.minecraftforge.client.model.data.EmptyModelData.INSTANCE);
+			ModelData.EMPTY,
+			renderType);
 	
 		matrix.popPose();
 	}
 
 	@Override
-	public void putQuadData(BlockAndTintGetter level, BlockState state, BlockPos pos, VertexConsumer buffer, PoseStack.Pose matrixEntry, BakedQuad quadIn,
+	public void putQuadData(BlockAndTintGetter level, BlockState state, BlockPos pos, VertexConsumer vertexConsumer, PoseStack.Pose pose, BakedQuad quad,
 		float tintA, float tintB, float tintC, float tintD, int brightness0, int brightness1, int brightness2, int brightness3, int combinedOverlayIn)
 	{
 		float r=1F;
 		float g=1F;
 		float b=1F;
-		if (quadIn.isTinted())
+		float alpha = ClientProxy.clientConfig.placementPreviewOpacity.get().floatValue();
+		if (quad.isTinted())
 		{
-			int i = this.blockColors.getColor(state, level, pos, quadIn.getTintIndex());
+			int i = this.blockColors.getColor(state, level, pos, quad.getTintIndex());
 			r = (i >> 16 & 255) / 255.0F;
 			g = (i >> 8 & 255) / 255.0F;
 			b = (i & 255) / 255.0F;
 		}
-		if (quadIn.isShade()) // better name: shouldApplyDiffuseLighting
-		{
-			float forgeLighting = net.minecraftforge.client.model.pipeline.LightUtil.diffuseLight(quadIn.getDirection());
-			r *= forgeLighting;
-			g *= forgeLighting;
-			b *= forgeLighting;
-		}
-
-		// use our method below instead of adding the quad in the usual manner
-		addTransparentQuad(matrixEntry, quadIn, new float[] { tintA, tintB, tintC, tintD }, r, g, b, new int[] { brightness0, brightness1, brightness2, brightness3 },
-			combinedOverlayIn, true, buffer);
-	}
-	
-	// as IVertexBuilder::addQuad except when we add the vertex, we add an opacity float instead of 1.0F
-	public static void addTransparentQuad(PoseStack.Pose matrixEntry, BakedQuad quad, float[] colorMuls, float r, float g, float b, int[] vertexLights,
-		int combinedOverlayIn, boolean mulColor, VertexConsumer buffer)
-	{
-		int[] vertexData = quad.getVertices();
-		Vec3i faceVector3i = quad.getDirection().getNormal();
-		Vector3f faceVector = new Vector3f(faceVector3i.getX(), faceVector3i.getY(), faceVector3i.getZ());
-		Matrix4f matrix = matrixEntry.pose();
-		faceVector.transform(matrixEntry.normal());
 		
-		int vertexDataEntries = vertexData.length / 8;
-
-		try (MemoryStack memorystack = MemoryStack.stackPush())
-		{
-			ByteBuffer bytebuffer = memorystack.malloc(DefaultVertexFormat.BLOCK.getVertexSize());
-			IntBuffer intbuffer = bytebuffer.asIntBuffer();
-
-			for (int vertexIndex = 0; vertexIndex < vertexDataEntries; ++vertexIndex)
-			{
-				((Buffer) intbuffer).clear();
-				intbuffer.put(vertexData, vertexIndex * 8, 8);
-				float x = bytebuffer.getFloat(0);
-				float y = bytebuffer.getFloat(4);
-				float z = bytebuffer.getFloat(8);
-				float red = colorMuls[vertexIndex] * r;
-				float green = colorMuls[vertexIndex] * g;
-				float blue = colorMuls[vertexIndex] * b;
-				
-				
-				if (mulColor)
-				{
-					float redMultiplier = (bytebuffer.get(12) & 255) / 255.0F;
-					float greenMultiplier = (bytebuffer.get(13) & 255) / 255.0F;
-					float blueMultiplier = (bytebuffer.get(14) & 255) / 255.0F;
-					red = redMultiplier * red;
-					green = greenMultiplier * green;
-					blue = blueMultiplier * blue;
-				}
-				
-				// this is the important part
-				float alpha = ClientProxy.clientConfig.placementPreviewOpacity.get().floatValue();
-
-				int light = buffer.applyBakedLighting(vertexLights[vertexIndex], bytebuffer);
-				float texU = bytebuffer.getFloat(16);
-				float texV = bytebuffer.getFloat(20);
-				Vector4f posVector = new Vector4f(x, y, z, 1.0F);
-				posVector.transform(matrix);
-				buffer.applyBakedNormals(faceVector, bytebuffer, matrixEntry.normal());
-				buffer.vertex(posVector.x(), posVector.y(), posVector.z(), red, green, blue, alpha, texU, texV,
-					combinedOverlayIn, light, faceVector.x(), faceVector.y(), faceVector.z());
-			}
-		}
+		// putBulkData with alpha value instead of 1
+		vertexConsumer.putBulkData(pose, quad, new float[]{tintA, tintB, tintC, tintD}, r, g, b, alpha, new int[]{brightness0, brightness1, brightness2, brightness3}, combinedOverlayIn, true);
 	}
 }